refactor(frontend): split engine contract into engine/ modules
Split pegainfer-frontend/src/engine.rs (1221 lines) into a 7-file
engine/ tree by cohesion: request / event / sink / kv / control /
handle, plus mod.rs. Pure move — no logic changes:

- private mod + flat pub use re-exports keep every
  pegainfer_frontend::engine::X path unchanged, so model crates,
  the vllm stack, and the server build untouched
- tests moved next to their subject (sink.rs x3, handle.rs x11)
- only mechanical adjustments: KvPrefix::rank private -> pub(crate);
  two cross-module rustdoc links crate-prefixed
- review follow-up: launch-time EngineLoadOptions/EpBackend live in
  handle.rs; request.rs holds only GenerateRequest
- sync the crate-layout block in docs/subsystems/frontend/

Verified: cargo check -p pegainfer-frontend --all-targets (0 warnings),
cargo test -p pegainfer-frontend (56/56), cargo check --workspace
--all-targets (0 errors), cargo fmt --check clean.
