+++ docs/index.md
@@ -157,7 +157,7 @@ Organized by domain (model line / subsystem / playbook / lesson) instead of by l
 
 | Path | TL;DR |
 | --- | --- |
-| `subsystems/frontend/frontend-architecture.md` | `pegainfer-frontend` owns everything north of the model schedulers: engine contract (ex `pegainfer-engine`), `vllm` protocol stack (ex `pegainfer-vllm-frontend`), `ModelLine` dispatch. All six model lines implement `ModelLine` in their own crates; `pegainfer-server` is a pure-dispatch bin. Next: prototype a `dynamo` in-process stack. |
+| `subsystems/frontend/frontend-architecture.md` | `pegainfer-frontend` owns everything north of the model schedulers: engine contract (ex `pegainfer-engine`, now the `engine/` module tree), `vllm` protocol stack (ex `pegainfer-vllm-frontend`), `ModelLine` dispatch. All six model lines implement `ModelLine` in their own crates; `pegainfer-server` is a pure-dispatch bin. Next: prototype a `dynamo` in-process stack. |
 | `subsystems/frontend/simulated-inference-engine.md` | CPU-only simulated model crate for vLLM/OpenAI frontend and `vllm bench serve` validation without CUDA, real model weights, or real-model performance claims. |
 | `subsystems/frontend/cpu-profiling-baseline.md` | Frontend CPU profiling baseline using `pegainfer-sim` with fixed TTFT=5ms/TPOT=12ms: 200 req / concurrency=16 shows ~150ms TTFT overhead (no dominant hotspot), heap allocation ~10%, stream polling ~7.5%, IPC ~1%; reproducible benchmark command and perf evidence documented. |
 | `subsystems/frontend/startup-time.md` | Qwen3-4B warm startup-to-ready: frontend tokenizer load runs concurrently with the engine load (HTTP still binds only after the engine registers); mmap teardown is paid at the end of load since #377; pinned-staging upload (2026-07) cuts warm ready 5.22s → 4.66s on sm_89, and the remaining floor is the engine's own post-load startup work. |
